fix(email): checkpoint-1 cleanup — shared test fakes, Google copy fix
- Lift _ScriptedConsole/_FakeAgent out of test_mailbox_onboarding_2469.py
  into tests/conftest.py so onboarding test modules share one fake
  instead of maintaining copies that can quietly drift apart.
- Fix a grammar regression in the client-credentials question: when both
  an ID and a secret are needed the copy now correctly says "them"/
  "they're" instead of "it"/"it's".
